requirements.txt
# Core dependencies
elasticsearch>=8.0.0,<9.0.0
google-cloud-pubsub>=2.0.0

# Optional: for production embedding
# transformers>=4.30.0
# torch>=2.0.0

# Optional: for caching
# redis>=4.0.0
